% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_bing}
\alias{link_bing}
\title{link_bing}
\usage{
link_bing(
  vector_of_gene_symbols,
  bing = "https://www.bing.com/search?q = ",
  prefix = "",
  suffix = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open,
  sleep = 0
)
}
\arguments{
\item{sleep}{}
}
\description{
link_bing
}
\examples{
link_bing("ACTB")
}
