% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_CGC}
\alias{link_CGC}
\title{link_CGC}
\usage{
link_CGC(vector_of_gene_symbols, writeOut = b.dbl.writeOut, Open = b.dbl.Open)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse CGC links (worms mutant database).
}
