% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_ensembl}
\alias{link_ensembl}
\title{link_ensembl}
\usage{
link_ensembl(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{writeOut}{Write links out to a file, Default: b.dbl.writeOut}

\item{Open}{Open a web browser with the search questy link, Default: b.dbl.Open}
}
\description{
Parse the latest ensembl (GRC38) links to your list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_ensembl(geneSymbols); link_ensembl(geneSymbols, Open = TRUE)
}
