% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_Franklin_clip2clip}
\alias{link_Franklin_clip2clip}
\title{link_Franklin_clip2clip}
\usage{
link_Franklin_clip2clip(
  coordinates = clipr::read_clip_tbl(header = F),
  searchQueryPrefix = "https://franklin.genoox.com/clinical-db/variant/snp/",
  as.ExcelLink = T
)
}
\arguments{
\item{coordinates}{A two-column dataframe of coordinates in the format (e.g., 5:35162876 C/T or 16:7164219 T/G) read from the clipboard. Default: clipboard content.}

\item{searchQueryPrefix}{The base URL for Genoox Franklin search query. Default: 'https://franklin.genoox.com/clinical-db/variant/snp/'.}

\item{as.ExcelLink}{A logical indicating whether to format the links as Excel links. Default: TRUE.}
}
\description{
Generate Franklin (Genoox) links from a list of coordinates.
}
\examples{
link_Franklin_clip2clip(coordinates = clipr::read_clip_tbl(header = F))

}
