% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_google}
\alias{link_google}
\title{link_google}
\usage{
link_google(
  vector_of_gene_symbols = c("TOP2A", "ACTB"),
  google = "http://www.google.com/search?as_q = ",
  prefix = "",
  suffix = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open,
  sleep = 0
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{google}{google search string, "http://www.google.com/search?as_q = "}

\item{prefix}{A search prefix string to add before each gene symbol, e.g. 'gene' or 'brain'; Default: ''}

\item{suffix}{A search suffix string to add after each gene symbol, e.g. 'gene' or 'brain'; Default: ''}

\item{writeOut}{Write links out to a file, Default: b.dbl.writeOut}

\item{Open}{Open a web browser with the search questy link, Default: b.dbl.Open}

\item{sleep}{How many seconds to wait between opening tow links? Increase so that its easier for your computer , Default: 1}
}
\description{
Parse google search query links to your list of gene symbols. Strings "prefix" and ""suffix" will be searched for together with each gene ("Human ID4 neurons"). See updated google and bing services in \href{https://vertesy.github.io/CodeAndRoll.R}{CodeAndRoll.R}.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_google(geneSymbols); link_google(geneSymbols, Open = TRUE)
}
