% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_MGI.JAX}
\alias{link_MGI.JAX}
\title{link_MGI.JAX}
\usage{
link_MGI.JAX(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols for which to generate MGI JAX links.}

\item{writeOut}{A logical indicating whether to write the generated links to a file, default: FALSE.}

\item{Open}{A logical indicating whether to open the generated links in a web browser, default: TRUE.}

\item{MGI_search_prefix}{The base URL for MGI JAX search.}

\item{MGI_search_suffix}{The suffix for MGI JAX search.}
}
\description{
Generate MGI JAX mouse genomics search query links for a list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_MGI.JAX(geneSymbols)
link_MGI.JAX(geneSymbols, Open = TRUE)

}
