% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_SNPedia_clip2clip}
\alias{link_SNPedia_clip2clip}
\title{link_SNPedia_clip2clip}
\usage{
link_SNPedia_clip2clip(
  rdIDs = clipr::read_clip_tbl(header = F),
  searchQueryPrefix = "https://www.snpedia.com/index.php/",
  as.ExcelLink = T,
  as.MarkDownLink = F
)
}
\arguments{
\item{rdIDs}{A list of rsIDs from an Excel column.}

\item{searchQueryPrefix}{The base URL for SNPedia search, default: 'https://www.snpedia.com/index.php/'.}

\item{as.ExcelLink}{A logical indicating whether to format the links as Excel links, default: TRUE.}

\item{as.MarkDownLink}{A logical indicating whether to format the links as Markdown links, default: FALSE.}
}
\description{
Generate SNPedia links from a list of rsIDs copied from an Excel column.
}
\examples{
link_SNPedia_clip2clip(rdIDs = clipr::read_clip_tbl(header=F))

}
