% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_String}
\alias{link_String}
\title{Parse STRING Database Links}
\usage{
link_String(
  vector_of_gene_symbols,
  organism = "human",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols.}

\item{organism}{A string indicating the organism. Can be 'mouse', 'human' or 'elegans'. Default is 'human'.}

\item{writeOut}{A logical indicating whether to write out the links. Default is b.dbl.writeOut.}

\item{Open}{A logical indicating whether to open the links. Default is b.dbl.Open.}
}
\description{
This function generates links to the STRING protein interaction database for a list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_String(geneSymbols); link_String(geneSymbols, Open = TRUE)

}
