% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_uniprot_mice}
\alias{link_uniprot_mice}
\title{Parse UNIPROT Links for Mice Genes}
\usage{
link_uniprot_mice(vector_of_gene_symbols, writeOut = F, Open = b.dbl.Open)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of mouse gene symbols.}

\item{writeOut}{A logical indicating whether to write out the links. Default is FALSE.}

\item{Open}{A logical indicating whether to open the links. Default is b.dbl.Open.}
}
\description{
This function generates UNIPROT links for a list of mouse gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_uniprot_mice(geneSymbols); link_uniprot_mice(geneSymbols, Open = TRUE)

}
