% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_wikipedia}
\alias{link_wikipedia}
\title{link_wikipedia}
\usage{
link_wikipedia(
  vector_of_gene_symbols = c("TOP2A", "ACTB"),
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{writeOut}{Write links out to a file, Default: b.dbl.writeOut}

\item{Open}{Open a web browser with the search questy link, Default: b.dbl.Open}
}
\description{
Parse wikipedia search query links to your list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_wikipedia(geneSymbols); link_wikipedia(geneSymbols, Open = TRUE)
}
