% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{qString}
\alias{qString}
\title{qString}
\usage{
qString(
  vector_of_gene_symbols = c("TOP2A", "ACTB"),
  organism = "human",
  writeOut = F,
  Open = TRUE
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{organism}{Organism , Default: 'human'}

\item{writeOut}{Write links out to a file, Default: F}

\item{Open}{Open a web browser with the search questy link, Default: TRUE}
}
\description{
Parse STRING protein interaction database links to your list of gene symbols. "organism" can be mouse, human or NA.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); qString(geneSymbols); qString(geneSymbols, Open = TRUE)
}
