% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_ensembl.grc37}
\alias{link_ensembl.grc37}
\title{Parse Ensembl GRC37 links}
\usage{
link_ensembl.grc37(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols.}

\item{writeOut}{A logical indicating whether to write out the links. Default is b.dbl.writeOut.}

\item{Open}{A logical indicating whether to open the links. Default is b.dbl.Open.}
}
\description{
This function generates Ensembl GRC37 links for a list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_ensembl.grc37(geneSymbols)
link_ensembl.grc37(geneSymbols, Open = TRUE)

}
