% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_ensembl}
\alias{link_ensembl}
\title{Generate Ensembl Links}
\usage{
link_ensembl(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A vector containing gene symbols to be linked.}

\item{writeOut}{A boolean flag indicating whether to write out the output or not. Default is the value of \code{b.dbl.writeOut}.}

\item{Open}{A boolean flag indicating whether to open the links. Default is the value of \code{b.dbl.Open}.}
}
\description{
This function generates the latest Ensembl (GRC38) links for a list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_ensembl(geneSymbols); link_ensembl(geneSymbols, Open = TRUE)

}
