% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_GeneCards}
\alias{link_GeneCards}
\title{GeneCards Link Generator}
\usage{
link_GeneCards(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector containing gene symbols.}

\item{writeOut}{A logical parameter determining if URLs should be written out to a bash script. Default is b.dbl.writeOut.}

\item{Open}{A logical parameter determining if URLs should be opened immediately. Default is b.dbl.Open.}
}
\description{
Generates GeneCards URLs for a given list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_GeneCards(geneSymbols)
link_GeneCards(geneSymbols, Open = TRUE)

}
