% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_google}
\alias{link_google}
\title{link_google}
\usage{
link_google(
  vector_of_gene_symbols,
  prefix = "",
  suffix = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open,
  sleep = 1
)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION, Default: ''}

\item{suffix}{PARAM_DESCRIPTION, Default: ''}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}

\item{sleep}{PARAM_DESCRIPTION, Default: 1}
}
\description{
Parse google search query links to your list of gene symbols. Strings "prefix" and ""suffix" will be searched for together with each gene ("Human ID4 neurons"). See updated google and bing services in \href{https://vertesy.github.io/CodeAndRoll.R}{CodeAndRoll.R}.
}
