% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_google}
\alias{link_google}
\title{link_google}
\usage{
link_google(
  vector_of_gene_symbols,
  google = "http://www.google.com/search?as_q = ",
  prefix = "",
  suffix = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open,
  sleep = 0
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols to generate Google search links.}

\item{google}{A Google search URL string, default: "http://www.google.com/search?as_q = ".}

\item{prefix}{A string to be added before each gene symbol in the search query, default: ''.}

\item{suffix}{A string to be added after each gene symbol in the search query, default: ''.}

\item{writeOut}{A boolean indicating whether to write the bash commands to file, default: b.dbl.writeOut.}

\item{Open}{A boolean indicating whether to open the search links, default: b.dbl.Open.}

\item{sleep}{The sleep interval in seconds between opening search links, default: 1.}
}
\description{
Parses Google search query links for a provided list of gene symbols.
The "prefix" and "suffix" will be searched for together with each gene (e.g., "Human ID4 neurons").
Uses google="http://www.google.com/search?as_q=".
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_google(geneSymbols); link_google(geneSymbols, Open = TRUE)

}
