% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_MGI.JAX}
\alias{link_MGI.JAX}
\title{link_MGI.JAX}
\usage{
link_MGI.JAX(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse MGI JAX mouse genomics links search query links to your list of gene symbols.
}
