% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_pubmed}
\alias{link_pubmed}
\title{link_pubmed}
\usage{
link_pubmed(
  vector_of_gene_symbols,
  additional_terms = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols for which links should be created.}

\item{additional_terms}{A character vector of additional terms to be included in the search query. Default is an empty string.}

\item{writeOut}{A boolean indicating whether to write out the commands or not. Default is determined by b.dbl.writeOut.}

\item{Open}{A boolean indicating whether to open the generated links or not. Default is determined by b.dbl.Open.}
}
\description{
Generates links to the PUBMED database based on a given list of gene symbols and additional search terms.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_pubmed(geneSymbols)
link_pubmed(geneSymbols, Open = TRUE)

}
