% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_pubmed}
\alias{link_pubmed}
\title{link_pubmed}
\usage{
link_pubmed(
  vector_of_gene_symbols,
  additional_terms = "",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{additional_terms}{PARAM_DESCRIPTION, Default: ''}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse PUBMED database links to your list of gene symbols. "additional_terms" can be any vector of strings that will be searched for together with each gene.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_pubmed(geneSymbols); link_pubmed(geneSymbols, Open = TRUE)
}
