% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_String}
\alias{link_String}
\title{link_String}
\usage{
link_String(
  vector_of_gene_symbols,
  organism = "human",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{organism}{PARAM_DESCRIPTION, Default: 'human'}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse STRING protein interaction database links to your list of gene symbols. "organism" can be mouse, human or NA.
}
