% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_String}
\alias{link_String}
\title{link_String}
\usage{
link_String(
  vector_of_gene_symbols = c("TOP2A", "ACTB"),
  organism = "human",
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{organism}{Organism, Default: 'human'}

\item{writeOut}{Write links out to a file, Default: b.dbl.writeOut}

\item{Open}{Open a web browser with the search questy link, Default: b.dbl.Open}
}
\description{
Parse STRING protein interaction database links to your list of gene symbols. "organism" can be mouse, human or NA.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_String(geneSymbols); link_String(geneSymbols, Open = TRUE)
}
