% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_uniprot_human}
\alias{link_uniprot_human}
\title{link_uniprot_human}
\usage{
link_uniprot_human(vector_of_gene_symbols, writeOut = F, Open = b.dbl.Open)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: F}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse the latest UNIPROT links to your list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_uniprot_human(geneSymbols); link_uniprot_human(geneSymbols, Open = TRUE)
}
