% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_uniprot_mice}
\alias{link_uniprot_mice}
\title{link_uniprot_mice}
\usage{
link_uniprot_mice(vector_of_gene_symbols, writeOut = F, Open = b.dbl.Open)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: F}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse the latest UNIPROT links to your list of gene symbols.
}
