% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_uniprot_mice}
\alias{link_uniprot_mice}
\title{link_uniprot_mice}
\usage{
link_uniprot_mice(vector_of_gene_symbols, writeOut = F, Open = b.dbl.Open)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols}

\item{writeOut}{Write links out to a file, Default: F}

\item{Open}{Open a web browser with the search questy link, Default: b.dbl.Open}
}
\description{
Parse the latest UNIPROT links to your list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_uniprot_mice(geneSymbols); link_uniprot_mice(geneSymbols, Open = TRUE)
}
