% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_wikipedia}
\alias{link_wikipedia}
\title{link_wikipedia}
\usage{
link_wikipedia(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols for which links should be created.}

\item{writeOut}{A boolean indicating whether to write out the commands or not. Default is determined by b.dbl.writeOut.}

\item{Open}{A boolean indicating whether to open the generated links or not. Default is determined by b.dbl.Open.}
}
\description{
Generates Wikipedia search query links based on a given list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_wikipedia(geneSymbols)
link_wikipedia(geneSymbols, Open = TRUE)

}
