% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_wormbase}
\alias{link_wormbase}
\title{link_wormbase}
\usage{
link_wormbase(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols for which to generate Wormbase links.}

\item{writeOut}{A logical indicating whether to write the generated links to a file, default: FALSE.}

\item{Open}{A logical indicating whether to open the generated links in a web browser, default: TRUE.}

\item{wormbase_search_prefix}{The base URL for Wormbase search.}
}
\description{
Generate Wormbase database links for a list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
link_wormbase(geneSymbols)
link_wormbase(geneSymbols, Open = TRUE)

}
