% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{link_wormbase}
\alias{link_wormbase}
\title{link_wormbase}
\usage{
link_wormbase(
  vector_of_gene_symbols,
  writeOut = b.dbl.writeOut,
  Open = b.dbl.Open
)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: b.dbl.writeOut}

\item{Open}{PARAM_DESCRIPTION, Default: b.dbl.Open}
}
\description{
Parse wormbase database links to your list of gene symbols..
}
