% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{openURLs.1by1}
\alias{openURLs.1by1}
\title{openURLs.1by1}
\usage{
openURLs.1by1(links, wait = 1)
}
\arguments{
\item{links}{PARAM_DESCRIPTION}

\item{wait}{PARAM_DESCRIPTION, Default: 1}
}
\description{
Open links with ~1 second delay. Necessary for slower computers and not to trigger the "robot" blocker of search engines (where you will pass your commands).
}
\examples{
geneSymbols = c('Sox2', 'Actb'); openURLs.1by1(geneSymbols)
}
