% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{qHGNC}
\alias{qHGNC}
\title{HGNC link generator and web lookup}
\usage{
qHGNC(
  vector_of_gene_symbols,
  HGNC_symbol_search = "https://www.genenames.org/tools/search/#!/?query=",
  writeOut = FALSE,
  Open = TRUE
)
}
\arguments{
\item{vector_of_gene_symbols}{A character vector of gene symbols to generate HGNC links.}

\item{HGNC_symbol_search}{A HGNC search URL string, default:
"https://www.genenames.org/tools/search/#!/?query=".}

\item{writeOut}{A boolean indicating whether to write the bash commands to file, default: FALSE.}

\item{Open}{A boolean indicating whether to open the HGNC links, default: TRUE.}
}
\description{
Parses HGNC links for a provided list of gene symbols.
}
\examples{
geneSymbols <- c("Sox2", "Actb")
qHGNC(geneSymbols)
qHGNC(geneSymbols, Open = FALSE)

}
