% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{qHGNC}
\alias{qHGNC}
\title{qHGNC}
\usage{
qHGNC(vector_of_gene_symbols, writeOut = FALSE, Open = TRUE)

qHGNC(vector_of_gene_symbols, writeOut = FALSE, Open = TRUE)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{writeOut}{PARAM_DESCRIPTION, Default: FALSE}

\item{Open}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Parse HGNC links to your list of gene symbols.

Parse HGNC links to your list of gene symbols.
}
\examples{
geneSymbols = c('Sox2', 'Actb'); link_HGNC(geneSymbols); link_HGNC(geneSymbols, Open = TRUE)
geneSymbols = c('Sox2', 'Actb'); qHGNC(geneSymbols); qHGNC(geneSymbols, Open = TRUE)
}
