% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseLinke.R
\name{qString}
\alias{qString}
\title{qString}
\usage{
qString(vector_of_gene_symbols, organism = "human", writeOut = F, Open = TRUE)
}
\arguments{
\item{vector_of_gene_symbols}{PARAM_DESCRIPTION}

\item{organism}{PARAM_DESCRIPTION, Default: 'human'}

\item{writeOut}{PARAM_DESCRIPTION, Default: F}

\item{Open}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Parse STRING protein interaction database links to your list of gene symbols. "organism" can be mouse, human or NA.
}
